Don't seed data with EF Core. Seeding data with relationships requires storing a Guid somewhere. This would likely need to be static, reserving useless memory for the life of the application.
Data seeding could be done in a SQL script instead.

FanVoltage.DTO.Edit:
 -Voltage should be a float

Chassis.Chassis:
 -Edit method should call base implementation

Part.Part:
 -Only expose edit method to descendents

Session.SessionController:
 -Fixing session management

SessionCookie:
 -Parsing the cookie was uneccesary

cc_purgetables:
 -Deleting tables was out-of-order for dependencies

+++ computer-configurator-api/FanVoltage/DTO/Edit.cs
@@ -3,7 +3,7 @@
     public class Edit
     {
         public Guid UUID { get; set; } = Guid.Empty;
-        public int Voltage { get; set; }
+        public float Voltage { get; set; }
 
         public IReadOnlyList<string> Validate() => new Validation(this).Errors;
     }
